requirements.txt
azure-common==1.1.28 ; python_version >= "3.10" and python_version < "4.0"
azure-core==1.25.1 ; python_version >= "3.10" and python_version < "4.0"
azure-identity==1.10.0 ; python_version >= "3.10" and python_version < "4.0"
azure-mgmt-core==1.3.2 ; python_version >= "3.10" and python_version < "4.0"
azure-mgmt-monitor==4.0.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.6.15.1 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4"
cryptography==38.0.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.3 ; python_version >= "3.10" and python_version < "4"
isodate==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
msal-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
msal==1.18.0 ; python_version >= "3.10" and python_version < "4.0"
msrest==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.23.3 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
pandas==1.4.4 ; python_version >= "3.10" and python_version < "4.0"
portalocker==2.5.1 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pyjwt[crypto]==2.4.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.2.1 ; python_version >= "3.10" and python_version < "4.0"
pywin32==304 ; python_version >= "3.10" and platform_system == "Windows" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.12 ; python_version >= "3.10" and python_version < "4"
